delete para categoria

+++ src/main/java/com/bookstore/bookstoreapi/repositories/LivroRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.bookstore.bookstoreapi.domain.Livro;
 
 @Repository
-public interface LivroRepository extends JpaRepository<Livro, Long> {
+public interface LivroRepository extends JpaRepository<Livro, Integer> {
     
 }
